crs/requirements.txt
Flask==2.3.2
mysql-connector-python==8.0.33
pandas==2.0.3
numpy==1.25.2
scikit-learn==1.3.0 # Para la IA, aunque sea un modelo de juguete, puede ser útil.
